+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `AssertsPromises` trait, useful for testing data loaders.
 
 ## [3.2.0] - 2020-08-03